add comprehensive numeric type support
- Signed integers: Int8, Int16, Int32, Int64
- Unsigned integers: UInt8, UInt16, UInt32, UInt64
- Floats: Float32, Float64
- Arbitrary precision: BigInt (using num-bigint)
- Fixed-point: Decimal (using rust_decimal)

Literal suffixes: 42i32, 100u16, 3.14f32, 42n, 99d

Conversion builtins: toInt8, toInt16, toInt32, toInt64,
toUInt8, toUInt16, toUInt32, toUInt64, toFloat32, toFloat64,
toBigInt

JIT extended to support all numeric types with type detection.
Polymorphic arithmetic dispatch at runtime.

+++ crates/vm/benches/vm_vs_runtime.rs
@@ -36,7 +36,7 @@ fn bench_vm() -> (i64, std::time::Duration) {
     let elapsed = start.elapsed();
 
     let value = match result {
-        Value::Int(v) => v,
+        Value::Int64(v) => v,
         _ => panic!("Expected Int"),
     };
 
@@ -69,7 +69,7 @@ fn bench_runtime() -> (i64, std::time::Duration) {
 
     let result = result.expect("Runtime error");
     let value = match result {
-        Some(GcValue::Int(v)) => v,
+        Some(GcValue::Int64(v)) => v,
         other => {
             eprintln!("Expected Int, got {:?}", other);
             eprintln!("Compiler functions: {:?}", compiler.get_all_functions().keys().collect::<Vec<_>>());
